Chat style changes, creature small fix

+++ creature.h
@@ -496,8 +496,8 @@ class Creature : virtual public Thing
 		uint32_t lastStepCost = 1;
 		uint32_t baseSpeed = 220;
 		int32_t varSpeed = 0;
-		int32_t health = 1000;
-		int32_t healthMax = 1000;
+		int64_t health = 1000;
+		int64_t healthMax = 1000;
 
 		Outfit_t currentOutfit;
 		Outfit_t defaultOutfit;
